added xml files of studytube and edubookers
Make python scripts that can also make the xml files for studytube and edubookers. Edited the Springest XML so there is a minimum of errors with the XSD. Need reply from Springest if I need to fix these errors or that they can skip these nodes.

XML file from Studytube is correct with the XSD, they can skip the description so the planning can be uploaded with the XML without changing the descriptions.

Edubookers does not have a XSD, but they are taking a look at it if it will work, still need a reply. They can also skip the description so the planning can be uploaded without changing the descriptions.

Next step: cleaning the code and maybe merge some files